src/requirements.txt
discord.py
beautifulsoup4
requests
PyPDF2
EbookLib

